[Azoth Xoox] Fixed narrowing cast.

+++ src/plugins/azoth/plugins/xoox/entrybase.cpp
@@ -944,7 +944,7 @@ namespace Xoox
 			variant = "";
 
 		const auto secsDiff = QDateTime::currentDateTimeUtc ().secsTo (thatTime);
-		Variant2SecsDiff_ [variant] = { secsDiff, iq.tzo () };
+		Variant2SecsDiff_ [variant] = { static_cast<int> (secsDiff), iq.tzo () };
 
 		emit entryGenerallyChanged ();
